Correction of some parts and addition of exercise 4 resolution

+++ src/conditional_statements/Exercise6.java
@@ -13,7 +13,7 @@ public class Exercise6 {
 		
 		String name;
 		int code;
-		double salary, newSalary;
+		float salary, newSalary;
 		
 		System.out.println("Olá! Como posso te chamar?");
 		name = read.next();
@@ -34,42 +34,49 @@ public class Exercise6 {
 		salary = read.nextFloat();
 		
 		switch (code) {
+		
 		case 1:
-			newSalary = salary + (salary * 0.10);
+			newSalary = salary + (salary * 0.10f);
 			System.out.println("Nome do colaborador: " + name);
 			System.out.println("Cargo: Gerente");
 			System.out.println("Salário: " + currencyFormatter.format(newSalary));
 			break;
+			
 		case 2:
-			newSalary = salary + (salary * 0.07);
+			newSalary = salary + (salary * 0.07f);
 			System.out.println("Nome do colaborador: " + name);
 			System.out.println("Cargo: Vendendor");
 			System.out.println("Salário: " + currencyFormatter.format(newSalary));
 			break;
+			
 		case 3:
-			newSalary = salary + (salary * 0.09);
+			newSalary = salary + (salary * 0.09f);
 			System.out.println("Nome do colaborador: " + name);
 			System.out.println("Cargo: Supervisor");
 			System.out.println("Salário: " + currencyFormatter.format(newSalary));
 			break;
+			
 		case 4:
-			newSalary = salary + (salary * 0.06);
+			newSalary = salary + (salary * 0.06f);
 			System.out.println("Nome do colaborador: " + name);
 			System.out.println("Cargo: Motorista");
 			System.out.println("Salário: " + currencyFormatter.format(newSalary));
 			break;
+			
 		case 5:
-			newSalary = salary + (salary * 0.05);
+			newSalary = salary + (salary * 0.05f);
 			System.out.println("Nome do colaborador: " + name);
 			System.out.println("Cargo: Estoquista");
 			System.out.println("Salário: " + currencyFormatter.format(newSalary));
 			break;
+			
 		case 6:
-			newSalary = salary + (salary * 0.08);
+			newSalary = salary + (salary * 0.08f);
 			System.out.println("Nome do colaborador: " + name);
 			System.out.println("Cargo: Técnico de TI");
 			System.out.println("Salário: " + currencyFormatter.format(newSalary));
 			break;
+			
 		default:
 			System.out.println(name + " digite o código de colaborador válido.");
 		}
